improve compatibility
Some units don't respond to RH and Ra queries used to fetch inside and
outside temperatures, along with some other queries.

This commit fixes the compatibility by trying to read the inside and
outside temperatures with both the F9 or RH/Ra queries. Since we don't
know which of these can fail, they are moved to a separate loop that
won't affect the ready state, but it still fills all the gaps that can
occur due to not responding to some query.